refactor: AbstractUploadCommand
